feat: Add registerApplyFunc and apply methods to ConfigManager
The commit adds two new methods, `registerApplyFunc` and `apply`, to the `ConfigManager` class. The `registerApplyFunc` method allows registering a function that will be called when applying unsaved configurations, and the `apply` method applies all the unsaved configurations by calling the registered functions. This change enhances the functionality of the `ConfigManager` class by providing a way to handle unsaved configurations.